Update Dealership.java
I added a readRecords() function that will read in the data (if any) from the cars.txt file and stores it into the ArrayList. This lets you add new car records after the old ones to be able to just overwrite the cars.txt file with the old and new data every time you run the program.