Add files via upload
Here are the files I started with for now. I decided to use our code from Assignment 2, because his was weird and ours was broken into simple chunks when displaying and storing data, I think its easier to work with ours for now since we have a lot to do. We can always change it later when we link the stuff. I created a "SearchCarGUI.java" for a search window. I'm not sure how to link frames in the Car layout yet, but this is just the bare bones stuff to get us started.