Update DealershipMenuGUI.java
I added the file stream reading and writing, but I don't know where to call these functions? I did readDatabase() in the public DealershipMenuGUI(), but I don't know where to place the writeDatabase(). It won't work in the main function file, and I can't figure out how to add it on close.